Added more columns to the hts_fact table

+++ xf-system/sp/drop_objects/sp_xf_system_drop_all_tables_in_schema.sql
@@ -1,26 +1,35 @@
 DELIMITER //
 
-DROP PROCEDURE IF EXISTS dbo.sp_xf_system_drop_all_tables_in_schema;
+DROP PROCEDURE IF EXISTS dbo.sp_xf_system_drop_all_tables_in_schema//
 
 CREATE PROCEDURE dbo.sp_xf_system_drop_all_tables_in_schema(
     IN database_name NVARCHAR(255)
 )
 BEGIN
 
-    SET session group_concat_max_len = 20000;
+    DECLARE tables_count INT;
 
-    SET @tbls = (SELECT GROUP_CONCAT(database_name, '.', TABLE_NAME SEPARATOR ', ')
-                 FROM information_schema.tables
-                 WHERE TABLE_TYPE = 'BASE TABLE'
-                   AND TABLE_SCHEMA = database_name);
+    SELECT COUNT(1) INTO tables_count FROM information_schema.tables
+    WHERE TABLE_TYPE = 'BASE TABLE' AND TABLE_SCHEMA = database_name;
 
-    SET @drop_tables = CONCAT('DROP TABLE IF EXISTS ', @tbls);
+    IF tables_count > 0 THEN
 
-    SET foreign_key_checks = 0; -- Remove check, so we don't have to drop tables in the correct order, nor whether they actually exist.
-    PREPARE drop_tbls FROM @drop_tables;
-    EXECUTE drop_tbls;
-    DEALLOCATE PREPARE drop_tbls;
-    SET foreign_key_checks = 1;
+        SET session group_concat_max_len = 20000;
+
+        SET @tbls = (SELECT GROUP_CONCAT(database_name, '.', TABLE_NAME SEPARATOR ', ')
+                     FROM information_schema.tables
+                     WHERE TABLE_TYPE = 'BASE TABLE'
+                       AND TABLE_SCHEMA = database_name);
+
+        SET @drop_tables = CONCAT('DROP TABLE IF EXISTS ', @tbls);
+
+        SET foreign_key_checks = 0; -- Remove check, so we don't have to drop tables in the correct order, nor whether they actually exist.
+        PREPARE drop_tbls FROM @drop_tables;
+        EXECUTE drop_tbls;
+        DEALLOCATE PREPARE drop_tbls;
+        SET foreign_key_checks = 1;
+
+    END IF;
 
 END//
